WIP use trait
