+++ src/Commands/AddCollection.php
@@ -18,7 +18,7 @@ use function Laravel\Prompts\text;
 
 class AddCollection extends Command
 {
-    use RunsInPlease, SharedFunctions;
+    use RunsInPlease, SharedFunctions, NeedsValidLicense;
 
     protected $name = 'statamic:peak:add-collection';
     protected $description = "Add a collection.";
@@ -43,11 +43,6 @@ class AddCollection extends Command
 
     public function handle()
     {
-        if (! static::$licensed) {
-            $this->info("You need a valid license to use this command.");
-            return;
-        }
-
         $this->collection_name = text(
             label: 'What should be the name for this collection?',
             placeholder: 'E.g. News',
