+++ src/Commands/NeedsValidLicense.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Studio1902\PeakCommands\Commands;
+
+use Facades\Statamic\Licensing\LicenseManager;
+
+trait NeedsValidLicense {
+
+    protected static $licensed = false;
+
+    public function __construct()
+    {
+        $addonLicense = LicenseManager::addons()->first(function ($addonLicense) {
+            return $addonLicense->addon()->id() === 'studio1902/statamic-peak-commands';
+        });
+
+        static::$licensed = $addonLicense && $addonLicense->valid();
+    }
+
+    public function handle() {
+        if (! static::$licensed) {
+            $this->info("You need a valid license to use this command.");
+            return;
+        }
+    }
+}
